requirements.txt
numpy~=1.26.4
torch~=2.3.1
scipy~=1.13.1
pandas~=2.2.2
pydantic~=2.7.3
pytest~=8.2.2
PyYAML~=6.0.1